Added stone id settings to biome configs.

+++ bukkit/src/com/khorn/terraincontrol/bukkit/CustomBiome.java
@@ -45,11 +45,12 @@ public class CustomBiome extends BiomeBase
     {
         this.D = config.BiomeHeight;
         this.E = config.BiomeVolatility;
-        this.A = config.SurfaceBlock;
-        this.B = config.GroundBlock;
+        this.A = (byte) config.SurfaceBlock;
+        this.B = (byte) config.GroundBlock;
         this.temperature = config.BiomeTemperature;
         this.humidity = config.BiomeWetness;
-        if(this.humidity == 0) {
+        if (this.humidity == 0)
+        {
             this.b(); // this.disableRain()
         }
 
@@ -76,10 +77,7 @@ public class CustomBiome extends BiomeBase
             } else
             {
                 // The .toLowerCase() is just a safeguard so that we get notified if this.y is no longer the biome name
-                TerrainControl.log(Level.WARNING, "Mob type {0} not found in {1}", new Object[]
-                {
-                    mobGroup.getMobName(), this.y.toLowerCase()
-                });
+                TerrainControl.log(Level.WARNING, "Mob type {0} not found in {1}", new Object[]{mobGroup.getMobName(), this.y.toLowerCase()});
             }
         }
     }
